# Much faster and less allocating azure client