--- a/.gitignore
+++ b/.gitignore
@@ -288,3 +288,4 @@ __pycache__/
 *.xsd.cs
 binaries/
 nugets/
+src/QueueBatch.Benchmarks/BenchmarkDotNet.Artifacts/