--- a/src/QueueBatch.Benchmarks/Program.cs
+++ b/src/QueueBatch.Benchmarks/Program.cs
@@ -1,0 +1,14 @@
+﻿using BenchmarkDotNet.Configs;
+using BenchmarkDotNet.Diagnosers;
+using BenchmarkDotNet.Running;
+
+namespace QueueBatch.Benchmarks
+{
+    class Program
+    {
+        static void Main(string[] args)
+        {
+            var summary = BenchmarkRunner.Run(typeof(Program).Assembly, DefaultConfig.Instance.With(new MemoryDiagnoser()));
+        }
+    }
+}