--- a/src/QueueBatch/Impl/BindingProvider.cs
+++ b/src/QueueBatch/Impl/BindingProvider.cs
@@ -6,6 +6,7 @@ using Microsoft.Azure.WebJobs.Host.Triggers;
 using Microsoft.Extensions.Logging;
 using Microsoft.WindowsAzure.Storage;
 using Microsoft.WindowsAzure.Storage.Queue;
+using QueueBatch.Impl.Queues;
 
 namespace QueueBatch.Impl
 {
@@ -31,15 +32,21 @@ namespace QueueBatch.Impl
 
             var queueName = attr.QueueName;
 
-            var queue = queues.GetQueueReference(queueName);
+            var messageQueue = queues.GetQueueReference(queueName);
             var poisonQueue = CreatePoisonQueue(queueName);
 
             await Task.WhenAll(
-                queue.CreateIfNotExistsAsync(),
+                messageQueue.CreateIfNotExistsAsync(),
                 poisonQueue.CreateIfNotExistsAsync()
             ).ConfigureAwait(false);
-            
-            return new TriggerBinding(context.Parameter, queue, poisonQueue, TimeSpan.FromSeconds(attr.MaxBackOffInSeconds), attr.ParallelGets, loggerFactory);
+
+            var cache = new HttpMessageHandlerExpiringCache(TimeSpan.FromSeconds(10));
+
+            var queue = attr.UseFasterQueues
+                ? new QueueFunctionLogic(SdkQueue.CreateFast(messageQueue, cache), SdkQueue.CreateFast(poisonQueue, cache))
+                : new QueueFunctionLogic(new SdkQueue(messageQueue), new SdkQueue(poisonQueue));
+
+            return new TriggerBinding(context.Parameter, queue, TimeSpan.FromSeconds(attr.MaxBackOffInSeconds), attr.ParallelGets, loggerFactory);
         }
 
         CloudQueue CreatePoisonQueue(string name)