--- a/src/QueueBatch/Impl/Queues/FastXmlAzureParser.cs
+++ b/src/QueueBatch/Impl/Queues/FastXmlAzureParser.cs
@@ -6,7 +6,7 @@ using System.Text;
 
 namespace QueueBatch.Impl.Queues
 {
-    class FastXmlAzureParser
+    static class FastXmlAzureParser
     {
         static readonly QueueMessageNodes Nodes = new QueueMessageNodes();