--- a/src/QueueBatch/Impl/Queues/IRetrievedMessages.cs
+++ b/src/QueueBatch/Impl/Queues/IRetrievedMessages.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Collections.Generic;
+
+namespace QueueBatch.Impl.Queues
+{
+    interface IRetrievedMessages : IDisposable
+    {
+        IEnumerable<Message> Messages { get; }
+    }
+}