--- a/src/QueueBatch/InternalsVisibileTo.cs
+++ b/src/QueueBatch/InternalsVisibileTo.cs
@@ -1,3 +1,4 @@
 ﻿using System.Runtime.CompilerServices;
 
-[assembly:InternalsVisibleTo("QueueBatch.Tests")] 
+[assembly:InternalsVisibleTo("QueueBatch.Tests")]
+[assembly:InternalsVisibleTo("QueueBatch.Benchmarks")]