--- a/src/QueueBatch/Message.cs
+++ b/src/QueueBatch/Message.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Diagnostics;
 
 namespace QueueBatch
 {