--- a/src/QueueBatch/QueueBatch.csproj
+++ b/src/QueueBatch/QueueBatch.csproj
@@ -19,5 +19,6 @@
     <PackageReference Include="System.Memory" Version="4.5.1" />
     <PackageReference Include="GitVersionTask" Version="4.0.0-beta0012" PrivateAssets="All" />
     <PackageReference Include="System.Net.Http" Version="4.3.3" />
+    <PackageReference Include="WindowsAzure.Storage" Version="9.3.1" />
   </ItemGroup>
 </Project>