… README.md
… src/QueueBatch.Tests/approved_api.txt
… src/QueueBatch/Impl/HttpMessageHandlerExpiringCache.cs
… src/QueueBatch/Impl/Listener.cs
… src/QueueBatch/Impl/Queues/FastCloudQueue.cs
… src/QueueBatch/Impl/Queues/QueueFunctionLogic.cs
… src/QueueBatch/Impl/Queues/Result.cs
… src/QueueBatch/Impl/Queues/SdkQueue.cs
… src/QueueBatch/Impl/StorageExtensions.cs
… src/QueueBatch/Impl/TriggerBinding.cs
… src/QueueBatch/MissingBits.cs
… src/QueueBatch/QueueBatchTriggerAttribute.cs