This PR introduces a huge performance boost by changing the Azure Storage Queue from the SDK to a custom one, based on `Span<T>` and `Memory<T>`. This makes the parsing and obtaining messages 20x faster.

### Plan of delivery

- [x] create a cache for `HttpMessageHandler` that will handle expiring and reference counting, similar to `IHttpClientFactory`
- [x] create a custom `GetMessages` parser, that will operate on `Span<byte>` instead of `XmlReader`
- [x] build `IQueue` that will provide a common abstraction for SDK-based and custom queue wrapper
- [x] allow opt-in with an attribute `[QueueBatchTrigger(UseFasterQueues = true)]`
- [ ] create end-to-end benchmark running a function over queue that has 1000_000 messages

### Details

The SDK parser uses `XmlReader` to parse messages. This allocates a lot and is extremely slow. With `Span` and `.IndexOf` it's possible to build much faster parser. Additionally, one can just obtain a segment of response with `Memory<byte>` that allows to allocate even less. Effectively, this PR, by pooling 4MB of memory for getting messages, allocates only tasks for requests and not much more. To see the parsing advantage, take a look at benchmarks below.

### Benchmarks

``` ini

BenchmarkDotNet=v0.11.1, OS=Windows 10.0.16299.611 (1709/FallCreatorsUpdate/Redstone3)
Intel Core i7-6700HQ CPU 2.60GHz (Skylake), 1 CPU, 8 logical and 4 physical cores
Frequency=2531257 Hz, Resolution=395.0606 ns, Timer=TSC
.NET Core SDK=2.1.302
  [Host]     : .NET Core 2.1.2 (CoreCLR 4.6.26628.05, CoreFX 4.6.26629.01), 64bit RyuJIT
  DefaultJob : .NET Core 2.1.2 (CoreCLR 4.6.26628.05, CoreFX 4.6.26629.01), 64bit RyuJIT


```
|                   Method |        Mean |      Error |     StdDev |  Gen 0 | Allocated |
|------------------------- |------------:|-----------:|-----------:|-------:|----------:|
|               Sdk_parser | 15,764.7 ns | 275.846 ns | 317.664 ns | 2.2888 |    7224 B |
| QueueBatch_custom_parser |    618.2 ns |   8.565 ns |   7.593 ns | 0.0124 |      40 B |